pyproject.toml
[project]
name = "random-forest-dmazilkin"
version = "0.1.0"
description = "This package contains Random Forest algorithm implementation."
authors = [
    {name = "Dmitrii Mazilkin",email = "dim.mazilkin@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "decision-tree-dmazilkin (>=0.1.0,<0.2.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
